Create the example.app.geode.security package containing classes and component extensions to the Apache Geode Security Framework.
Define the User and Role security model classes to represent users of the system and the roles the user has been assigned.

Define the SecurityRepository interface representing a Data Access Object (DAO) used to access security configuraiton meta-data stored in an external data source.

Extend the SecurityRepository interface with CachingSecurityRepository abstract class with support for caching security configuration meta-data in-memory.

Provide the JdbcSecurityRepository implemenation of the SecurityRepository interface to access security configuration meta-data stored in a JDBC DataSource.

Provide the SimpleSecurityManager provider implementation of the SecurityManager interface, which uses the SecurityRepository to access security configuration meta-data to enforce secure access to Apache Geode.

Add DDL and DML SQL scripts to define the security configuration meta-data data model in Hsqldb.